Completata la versione 2.
- Semplificata la gestione dei FolderID
- Risolto problema di verifica di Media già presenti

+++ src/main/SystemController.java
@@ -145,7 +145,7 @@ public interface SystemController {
      * @param folderID The ID associated to the folder to be checked.
      * @return A boolean value: {@code true} if the given folder has children, {@code false} otherwise.
      */
-    boolean folderHasChildren(long folderID);
+    boolean folderHasChildren(int folderID);
 
     /**
      * Searches the file system for the folder with the given ID and returns its children in the format of a {@code
@@ -154,14 +154,14 @@ public interface SystemController {
      * @param parentID The ID of the desired parent folder.
      * @return A {@code String} containing all sub-folders of the desired parent folder.
      */
-    String getSubFolders(long parentID);
+    String getSubFolders(int parentID);
 
     /**
      * Getter for the ROOT folder ID.
      *
      * @return The ID's {@code Long} value.
      */
-    long getRootID();
+    int getRootID();
 
     /**
      * Returns the path of the folder associated with the given ID.
@@ -169,7 +169,7 @@ public interface SystemController {
      * @param folderID The folder ID.
      * @return A {@code String} containing the full path of that folder.
      */
-    String getPathToString(long folderID);
+    String getPathToString(int folderID);
 
     /**
      * Returns the contents of the folder that matches the given path.
